Fixed a batch of circular file dependencies
