+++ src/enums/dwn-interface-method.ts
@@ -1,0 +1,20 @@
+export enum DwnInterfaceName {
+  Events = 'Events',
+  Messages = 'Messages',
+  Permissions = 'Permissions',
+  Protocols = 'Protocols',
+  Records = 'Records'
+}
+
+export enum DwnMethodName {
+  Configure = 'Configure',
+  Create = 'Create',
+  Get = 'Get',
+  Grant = 'Grant',
+  Query = 'Query',
+  Read = 'Read',
+  Request = 'Request',
+  Revoke = 'Revoke',
+  Write = 'Write',
+  Delete = 'Delete'
+}
