+++ src/types/event-types.ts
@@ -1,7 +1,7 @@
 import type { Event } from './event-log.js';
 import type { GenericMessageReply } from '../core/message-reply.js';
 import type { AuthorizationModel, GenericMessage } from './message-types.js';
-import type { DwnInterfaceName, DwnMethodName } from '../core/message.js';
+import type { DwnInterfaceName, DwnMethodName } from '../enums/dwn-interface-method.js';
 
 export type EventsGetDescriptor = {
   interface : DwnInterfaceName.Events;
