+++ src/types/messages-types.ts
@@ -1,6 +1,6 @@
 import type { GenericMessageReply } from '../core/message-reply.js';
 import type { AuthorizationModel, GenericMessage } from './message-types.js';
-import type { DwnInterfaceName, DwnMethodName } from '../core/message.js';
+import type { DwnInterfaceName, DwnMethodName } from '../enums/dwn-interface-method.js';
 
 export type MessagesGetDescriptor = {
   interface : DwnInterfaceName.Messages;
