+++ src/types/records-types.ts
@@ -7,7 +7,7 @@ import type { PublicJwk } from './jose-types.js';
 import type { RangeFilter } from './message-types.js';
 import type { Readable } from 'readable-stream';
 import type { AuthorizationModel, GenericMessage, GenericSignaturePayload, Pagination } from './message-types.js';
-import type { DwnInterfaceName, DwnMethodName } from '../core/message.js';
+import type { DwnInterfaceName, DwnMethodName } from '../enums/dwn-interface-method.js';
 
 export type RecordsWriteDescriptor = {
   interface: DwnInterfaceName.Records;
