+++ tests/interfaces/permissions-grant.spec.ts
@@ -4,13 +4,13 @@ import type { CreateFromPermissionsRequestOverrides } from '../../src/interfaces
 import type { PermissionScope } from '../../src/index.js';
 import type { RecordsPermissionScope } from '../../src/types/permissions-types.js';
 
+import { Message } from '../../src/core/message.js';
 import { PermissionsConditionPublication } from '../../src/types/permissions-types.js';
 import { PermissionsGrant } from '../../src/interfaces/permissions-grant.js';
 import { Secp256k1 } from '../../src/utils/secp256k1.js';
 import { TestDataGenerator } from '../utils/test-data-generator.js';
 import { Time } from '../../src/utils/time.js';
-import { DidKeyResolver, DwnErrorCode, Jws, PrivateKeySigner } from '../../src/index.js';
-import { DwnInterfaceName, DwnMethodName, Message } from '../../src/core/message.js';
+import { DidKeyResolver, DwnErrorCode, DwnInterfaceName, DwnMethodName, Jws, PrivateKeySigner } from '../../src/index.js';
 
 describe('PermissionsGrant', () => {
   describe('create()', async () => {
